adding tests to the update requesters. changing the distance intervals from int to float - might break existing implementations!

+++ core/src/main/java/com/novoda/location/LocatorSettings.java
@@ -23,9 +23,9 @@ public class LocatorSettings {
     private boolean useGps = Constants.USE_GPS;
     private boolean updateOnLocationChange = Constants.REFRESH_DATA_ON_LOCATION_CHANGED;
     private long updatesInterval = Constants.UPDATES_MAX_TIME;
-    private int updatesDistance = Constants.UPDATES_MAX_DISTANCE;
+    private float updatesDistance = Constants.UPDATES_MAX_DISTANCE;
     private long passiveUpdatesInterval = Constants.DEFAULT_INTERVAL_PASSIVE;
-    private int passiveUpdatesDistance = Constants.DEFAULT_DISTANCE_PASSIVE;
+    private float passiveUpdatesDistance = Constants.DEFAULT_DISTANCE_PASSIVE;
     private boolean enablePassiveUpdates = Constants.ENABLE_PASSIVE_UPDATES;
 
     public LocatorSettings(String packageName, String updateAction) {
@@ -73,11 +73,11 @@ public class LocatorSettings {
         this.updatesInterval = updatesInterval;
     }
 
-    public int getUpdatesDistance() {
+    public float getUpdatesDistance() {
         return updatesDistance;
     }
 
-    public void setUpdatesDistance(int updatesDistance) {
+    public void setUpdatesDistance(float updatesDistance) {
         this.updatesDistance = updatesDistance;
     }
 
@@ -89,11 +89,11 @@ public class LocatorSettings {
         this.passiveUpdatesInterval = passiveUpdatesInterval;
     }
 
-    public int getPassiveUpdatesDistance() {
+    public float getPassiveUpdatesDistance() {
         return passiveUpdatesDistance;
     }
 
-    public void setPassiveUpdatesDistance(int passiveUpdatesDistance) {
+    public void setPassiveUpdatesDistance(float passiveUpdatesDistance) {
         this.passiveUpdatesDistance = passiveUpdatesDistance;
     }
